requirements.txt
# Python 3.7.4
ConfigArgParse
h5py
zarr
#torch
#torchvision
batchgenerators
pytorch-lightning
neptune-client
visdom
nibabel
joblib
tqdm
psutil
python-dotenv
click
nilearn
hydra-core # --upgrade
# pip install -e ./